requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.30
pg8000==1.31.2
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3